+++ oceanicospy/analysis/temporal.py
@@ -6,7 +6,7 @@ from scipy.signal import resample,detrend
 from ..utils import wave_props
 
 class WaveTemporalAnalyzer:
-    def __init__(self,measured_signal,sampling_data,surface_level_column='eta[m]'): 
+    def __init__(self,measured_signal,sampling_data,surface_level_column='eta[m]',zero_centered=False): 
         """
         Initializes the analysis object with measurement signal and sampling data.
 
@@ -22,6 +22,8 @@ class WaveTemporalAnalyzer:
                 - ``burst_length_s`` (float): Duration of each burst in seconds.
         surface_level_column : str, optional
             The name of the column in measured_signal that contains surface level data (default is ``eta[m]``).
+        zero_centered: Boolean
+            Indicate if the measured signal is centered in zero, if not, measured signal is centered at zero (default is ``False``).
 
         Notes
         -----
@@ -34,6 +36,7 @@ class WaveTemporalAnalyzer:
         self.sampling_data = sampling_data
         self.burst_length_s = self.sampling_data['burst_length_s']
         self.surface_level_column = surface_level_column
+        self.zero_centered = zero_centered
 
     def _check_burst_length(self,burst_series):
         """Verify that the burst has the expected number of samples based on the sampling frequency and burst length.
@@ -131,8 +134,7 @@ class WaveTemporalAnalyzer:
         T = np.array(T)
 
         # Determine the wavenumber based on the dispersion relation
-        # TODO: the total or anchoring depth has to be used.
-        L = np.array([wave_props.wavelength(t, anchoring_depth) for t in T], dtype=np.float64) 
+        L = np.array([wave_props.wavelength(t, anchoring_depth+sensor_height) for t in T], dtype=np.float64) 
         k = 2.*np.pi/L
 
         # Computing non-adaptive transference factor Kp
@@ -173,9 +175,11 @@ class WaveTemporalAnalyzer:
         for i in self.measured_signal['burstId'].unique():
             burst_signal = self.measured_signal[self.measured_signal['burstId'] == i]
 
-            # TODO: validate this step, detrend option is already given in the observations. What if the data is coming from another source?
-            burst_signal_detrended = burst_signal.iloc[:,:-1].apply(lambda x: detrend(x,type='constant'), axis=0)
-            burst_signal_detrended[self.measured_signal.columns[-1]] = burst_signal.iloc[:, -1]
+            if self.zero_centered:
+                burst_signal_detrended = burst_signal.copy()
+            else:
+                burst_signal_detrended = burst_signal.iloc[:,:-1].apply(lambda x: detrend(x, type='constant'), axis=0)
+                burst_signal_detrended[self.measured_signal.columns[-1]] = burst_signal.iloc[:, -1]
 
             H_top_third, Hmax, Tmean, Lmean = self.apply_zero_upcrossing_burst(burst_signal_detrended[self.surface_level_column].values,
                                     self.sampling_data['anchoring_depth'], self.sampling_data['sensor_height'])